requirements.txt
Django
djangorestframework
requests
psycopg2-binary
django-environ
